feat(code cleanup): remove from main.go
